+++ demo/src/app/disqus/disqus.component.ts
@@ -1,16 +1,7 @@
-import {
-  Component,
-  Input,
-  Output,
-  OnChanges,
-  OnDestroy,
-  ChangeDetectionStrategy,
-  SimpleChanges,
-  Renderer2,
-  ElementRef,
-  EventEmitter
-} from '@angular/core';
+import { Component, Input, Output, OnChanges, OnDestroy, ChangeDetectionStrategy, Renderer2, ElementRef, EventEmitter } from '@angular/core';
 import { DisqusService } from './disqus.service';
+
+declare const window: any;
 
 @Component({
   selector: 'disqus',
@@ -20,86 +11,56 @@ import { DisqusService } from './disqus.service';
 
 export class DisqusComponent implements OnChanges, OnDestroy {
 
-  /** DISQUS Account Name */
-  @Input() shortname: string;
-
-  @Input() identifier: string;
   @Input() url: string;
-  @Input() categoryId: string;
-  @Input() language: string;
+  @Input() identifier: string;
   @Input() title: string;
+  @Input() category: string;
+  @Input() language: string;
 
   /** Track Comments */
   @Output() comment = new EventEmitter<any>();
 
-  constructor(private renderer: Renderer2, private el: ElementRef, private dService: DisqusService) {
-  }
+  constructor(private renderer: Renderer2, private el: ElementRef, private dService: DisqusService) { }
 
-  ngOnChanges(changes: SimpleChanges) {
+  ngOnChanges() {
 
-    if (!this.dService.disqus) {
+    if (!window.DISQUS) {
       this.addDisqusScript();
     } else {
-      const idChange = changes['identifier'];
-      const urlChange = changes['url'];
-      const catChange = changes['categoryId'];
-      const titleChange = changes['title'];
-      const langChange = changes['language'];
-
-      let isResetNeeded = false;
-
-      if (idChange && idChange.currentValue !== idChange.previousValue) {
-        isResetNeeded = true;
-      }
-      if (urlChange && urlChange.currentValue !== urlChange.previousValue) {
-        isResetNeeded = true;
-      }
-      if (catChange && catChange.currentValue !== catChange.previousValue) {
-        isResetNeeded = true;
-      }
-      if (titleChange && titleChange.currentValue !== titleChange.previousValue) {
-        isResetNeeded = true;
-      }
-      if (langChange && langChange.currentValue !== langChange.previousValue) {
-        isResetNeeded = true;
-      }
-
-      if (isResetNeeded) {
-        this.reset();
-      }
+      this.reset();
     }
   }
 
   addDisqusScript() {
 
-    /** Set disqus config */
-    this.dService.disqusConfig = this.getConfig();
+    /** Set DISQUS config */
+    window.disqus_config = this.getConfig();
 
     /** Add DISQUS script */
     const disqusScript = this.renderer.createElement('script');
-    disqusScript.src = `//${this.shortname}.disqus.com/embed.js`;
+    disqusScript.src = `//${this.dService.shortname}.disqus.com/embed.js`;
     disqusScript.async = true;
     disqusScript.type = 'text/javascript';
     this.renderer.setAttribute(disqusScript, 'data-timestamp', new Date().getTime().toString());
     this.renderer.appendChild(this.el.nativeElement, disqusScript);
   }
 
-  /** Reset disqus with new inputs. */
+  /** Reset DISQUS with the new inputs */
   reset() {
-    this.dService.disqus.reset({
+    window.DISQUS.reset({
       reload: true,
       config: this.getConfig()
     });
   }
 
-  /** Get disqus settings from inputs */
+  /** Get DISQUS settings from inputs */
   getConfig() {
     const self = this;
     return function () {
       this.page.identifier = self.identifier;
-      this.page.url = self.dService.validateUrl(self.url);
+      this.page.url = self.validateUrl(self.url);
       this.page.title = self.title;
-      this.category_id = self.categoryId;
+      this.category_id = self.category;
       this.language = self.language;
 
       /* Available callbacks are afterRender, onInit, onNewComment, onPaginate, onReady, preData, preInit, preReset */
@@ -109,13 +70,23 @@ export class DisqusComponent implements OnChanges, OnDestroy {
     };
   }
 
-  ngOnDestroy() {
-    if (this.dService.window) {
-      this.dService.window.DISQUS = undefined;
-      this.dService.window.disqusConfig = undefined;
-    } else {
-      (<any>global).DISQUS = undefined;
-      (<any>global).disqusConfig = undefined;
+  validateUrl(url: string) {
+    /** Validate URL input */
+    if (url) {
+      const r = /(http|https):\/\/(\w+:{0,1}\w*@)?(\S+)(:[0-9]+)?(\/|\/([\w#!:.?+=&%@!\-\/]))?/;
+
+      if (r.test(url)) {
+        return url;
+      } else {
+        console.warn('[Disqus]: Invalid URL, return undefined');
+      }
     }
+    /** DISQUS will fallback to "Window.location.href" when URL is undefined */
+    return undefined;
+  }
+
+  ngOnDestroy() {
+    window.DISQUS = undefined;
+    window.disqus_config = undefined;
   }
 }